added methods to store, retrieve and delete phone numbers in JSON
Name and phone number with unique id is stored in JSON format, and it can also be retrieved and deleted from it and the live changes can be seen on the JSON server. Improved existing code by adding methods to store, retrieve, and delete phone numbers in JSON, instead of using local storage to persist the data.